+++ apps/simpleclient_windows.c
@@ -196,7 +196,7 @@ discovery(const char *anchor, const char *uri, oc_string_array_t types,
   (void)interfaces;
   (void)bm;
   int i;
-  int uri_len = strlen(uri);
+  size_t uri_len = strlen(uri);
   uri_len = (uri_len >= MAX_URI_LENGTH) ? MAX_URI_LENGTH - 1 : uri_len;
   PRINT("\n\nDISCOVERYCB %s %s %d\n\n", anchor, uri,
         oc_string_array_get_allocated_size(types));
