+++ include/oc_collection.h
@@ -51,7 +51,7 @@ oc_collection_t *oc_collection_alloc(void);
 void oc_collection_free(oc_collection_t *collection);
 
 oc_collection_t *oc_get_collection_by_uri(const char *uri_path,
-                                          int uri_path_len, int device);
+                                          size_t uri_path_len, int device);
 oc_collection_t *oc_collection_get_all(void);
 oc_link_t *oc_get_link_by_uri(oc_collection_t *collection, const char *uri_path, int uri_path_len);
